D3ASIM-3348: Refactoring mqtt and websocket subscriber to live_data package.
