+++ d3a_api_client/constants.py
@@ -8,4 +8,4 @@ SETUP_FILE_PATH = None
 
 DEFAULT_DOMAIN_NAME = 'https://d3aweb.gridsingularity.com'
 DEFAULT_WEBSOCKET_DOMAIN = 'wss://d3aweb.gridsingularity.com/external-ws'
-CONSUMER_WEBSOCKET_DOMAIN_NAME = 'ws://localhost:4000'
+CUSTOMER_WEBSOCKET_DOMAIN_NAME = 'ws://localhost:4000'
