+++ integration_tests/build_test_containers.sh
@@ -14,7 +14,7 @@ if [[ "$(docker images -q ${D3A_IMAGE_TAG} 2> /dev/null)" == "" ]]; then
   echo "Building d3a image ..." && \
   rm -rf tests/d3a && \
   cd tests/ && \
-  git clone -b feature/D3ASIM-3348 https://github.com/gridsingularity/d3a.git && \
+  git clone https://github.com/gridsingularity/d3a.git && \
   cd d3a && \
   docker build -t ${D3A_IMAGE_TAG} . && \
   cd ../ && \
