+++ live_data/.envvars
@@ -3,12 +3,12 @@
 #     set -a ; source .envvars ; set +a
 #
 # in order to set up the API client and MQTT connections
-
 MQTT_USERNAME=""
 MQTT_PASSWORD=""
 MQTT_DOMAIN_NAME=""
 API_CLIENT_USERNAME=""
 API_CLIENT_PASSWORD=""
-API_CLIENT_DOMAIN_NAME="http://localhost:8000"
-API_CLIENT_WEBSOCKET_DOMAIN_NAME="ws://localhost:8000/external-ws"
+API_CLIENT_DOMAIN_NAME=""
+API_CLIENT_SIM_API_DOMAIN_NAME=""
+API_CLIENT_WEBSOCKET_DOMAIN_NAME=""
 REDIS_URL=""
