+++ live_data/mqtt_subscriber/oli_broker.py
@@ -4,7 +4,7 @@ import logging
 import traceback
 import paho.mqtt.client as mqtt
 from d3a_api_client.rest_device import RestDeviceClient
-from mqtt_subscriber import allowed_devices_name_mapping, generate_topic_api_client_args_mapping
+from live_data.mqtt_subscriber import allowed_devices_name_mapping, generate_topic_api_client_args_mapping
 from time import time
 
 
@@ -48,6 +48,7 @@ class MQTTConnection:
 
             # Transmit power values to CN
             for api_args in self.topic_api_client_dict[msg.topic]:
+                print(f"api_args: {api_args}")
                 RestDeviceClient(**api_args).set_energy_forecast(energy, do_not_wait=True)
 
         except Exception as e:
