+++ live_data/websocket_subscriber/__init__.py
@@ -1,0 +1,6 @@
+from live_data.websocket_subscriber.live_data_consumer import LiveDataConsumer
+
+
+def create_live_data_consumer():
+    print(f"create_live_data_consumer")
+    live_data = LiveDataConsumer()
