+++ live_data/websocket_subscriber/live_data_consumer.py
@@ -6,29 +6,19 @@ import asyncio
 
 from concurrent.futures.thread import ThreadPoolExecutor
 
-from d3a_api_client.utils import logging_decorator, domain_name_from_env, \
-    consumer_websocket_domain_name_from_env
+from d3a_api_client.utils import consumer_websocket_domain_name_from_env
 from d3a_api_client.websocket_device import WebsocketMessageReceiver
 from d3a_api_client.constants import MAX_WORKER_THREADS
 from d3a_api_client.websocket_device import retry_coroutine
-from d3a_api_client.utils import RestCommunicationMixin
-from d3a_api_client.utils import retrieve_jwt_key_from_server
+from live_data import LiveDataRest
 
 
 class LiveDataWebsocketMessageReceiver(WebsocketMessageReceiver):
-    def __init__(self, rest_client):
-        super().__init__(rest_client)
-
-    def _handle_event_message(self, message):
-        if message["event"] == "live_data":
-            self.client.set_energy_forecast(**message['data'])
-        else:
-            logging.error(f"Received message with unknown event type: {message}")
 
     def received_message(self, message):
         try:
-            if "event" in message:
-                self._handle_event_message(message)
+            if "event" in message and message["event"] == "live_data":
+                self.client.set_energy_forecast(**message['data'])
         except Exception as e:
             logging.error(f"Error while processing incoming message {message}. Exception {e}.\n"
                           f"{traceback.format_exc()}")
@@ -51,13 +41,11 @@ class LiveDataWebsocketThread(threading.Thread):
         event_loop.close()
 
 
-class LiveData(RestCommunicationMixin):
+class LiveDataConsumer(LiveDataRest):
 
     def __init__(self):
-        self.domain_name = domain_name_from_env
+        super().__init__()
         self.websockets_domain_name = consumer_websocket_domain_name_from_env
-        self.jwt_token = retrieve_jwt_key_from_server(self.domain_name)
-        self._create_jwt_refresh_timer(self.domain_name)
         self.start_websocket_connection()
 
     def start_websocket_connection(self):
@@ -68,9 +56,3 @@ class LiveData(RestCommunicationMixin):
         self.websocket_thread.start()
         self.callback_thread = ThreadPoolExecutor(max_workers=MAX_WORKER_THREADS)
 
-    @logging_decorator('set_energy_forecast')
-    def set_energy_forecast(self, *args, **kwargs):
-        self.simulation_id = kwargs['simulation_id']
-        self.device_id = kwargs['area_uuid']
-        transaction_id, posted = self._post_request('set_energy_forecast',
-                                                    {"energy_forecast": kwargs['energy_wh']})
